requirements.txt
pytest>=7.0.0
allure-pytest>=2.9.0
pytest-mock>=3.0.0